Analysis of Large Datasets to Identify Neonatal Morbidity Risks and Improve Family Health via Genetic Counseling
The Center for Disease Control (CDC) made some important recommendations as regards improving pre-conception health in the year 2006. The goal which is summarily to improve the health of women during pregnancy and the health of newborns in The United States. The CDC recommended increasing men's and women's pre-conception health knowledge, attitudes, and practices; another is to ensure that all American women of reproductive age receive pre-conception care services that will allow them to begin pregnancies in the best possible health. Given these benefits since over 18 years ago, the count of newborns without morbidity continues to decrease.
The objective of this project is to analyze risk of Hemolytic and Perinatal Jaundice by race and ethnicity; and to investigate the knowledge available to public, identify the factors majorly contributing to morbidity in newborns especially genetic incompatibility and provide recommendations for government and stakeholders on family health. The project also sought to identify the spread of knowledge about genotype, blood group and Rhesus factor; and more importantly Sickle Cell Disease. An application was developed to help the public calculate their genotypes.